Moved everything to PlayerActivity to get release out quickly (...+4)
Changelog:
- Removed all unused and static values from the main layout
- Moved all music player code to PlayerActivity:
  - Moved music playback controls and music playlist managing from MusicPlayerService and MainActivity into this
  - Moved GUI handling and started using ButterKnife for views
  - Did *not* import messaging utilities from MainActivity since this is all one class

Note:
I had to release a working version of this application for a school project, so I had to rush a rewrite into a week.
I opted to only rewrite part of the code, and make everything be managed by one Activity so that async stuff should have been less.