Added Preferences for API25+ devices (...+5)
Changelog:
- Added SettingsActivity:
    - Added preferences.xml holding the layout for the preferences
    - Added controls over specific debug preferences in SettingsActivity
    - Added string resources for both english and italian
- Added launch intent for SettingsActivity in AndroidManifest
- Added PrefsUtils class to help get and save preferences

Note:
Preferences are currently only available on Android Nougat 7.0 and above (API25+) due to the launch intent chosen for SettingsActivity.
To access the preferences the user will need to open the system Settings app and navigate through the list of installed apps to Music09.
Thanks to an API25 addition there will be a grind next to the app icon and name, usually at the right/end of the screen which, once tapped, will directly launch SettingsActivity.
There are currently no plans to integrate settings nicely into the app for devices older than Android Nougat 7.0 but there are plans to make it easier to access SettingsActivity from within the app on the other devices.