Updated Android Studio to 2.3 Stable
Note: It's suggested that Android 2.3 Stable is used from now on. This project will continue to follow the Canary channel, but will be backwards compatible to the last stable version release. Android Studio 2.3 will be, from now on, the minimum version for which this project will be considered compatible with.